Minor changes to project structure

+++ infolab-XY/src/XYstring.cpp
@@ -26,10 +26,10 @@ ull XYstring::iterationPivot(ull pos)
 	int power = 0;
 	power = static_cast<int>(trunc(log2(pos)));
 
-	ull middlePositon = pow(2, power);
+	ull middlePositon = static_cast<ull>(pow(2, power));
 
 	if (pos < middlePositon || power == ULL_SIZE)
-		middlePositon = pow(2, power - 1);
+		middlePositon = static_cast<ull>(pow(2, power - 1));
 
 	return middlePositon;
 }
